requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.6.1
watchfiles==0.21.0
feedparser==6.0.10
psycopg2-binary==2.9.9
apscheduler==3.10.4
